Fixed Balance calculation

+++ PantAPI/PantAPI/Models/BalanceResultModel.cs
@@ -7,7 +7,7 @@ namespace PantAPI.Models
 {
     public class BalanceResultModel
     {
-        public decimal Balance { get; set; }
+        public double Balance { get; set; }
         public List<BagInfo> Details {get; set;}
     }
 
@@ -15,8 +15,8 @@ namespace PantAPI.Models
     {
         public DateTime Registred { get; set; }
         public DateTime? Received { get; set; }
-        public decimal Value { get; set; }
-        public decimal? Weight { get; set; }
+        public double Value { get; set; }
+        public double? Weight { get; set; }
     }
 
 }
